Runtime/RTEfi*: Start of a VFS module to access a EFI variable store stored in the NVRAM file to allow changing the firmware config without having the VM running, bugref:9580 [build fix]
svn:sync-xref-src-repo-rev: r145559

+++ src/VBox/Runtime/common/efi/efivarstorevfs.cpp
@@ -1,4 +1,4 @@
-/* $Id: efivarstorevfs.cpp 90063 2021-07-06 12:50:55Z alexander.eichner@oracle.com $ */
+/* $Id: efivarstorevfs.cpp 90064 2021-07-06 12:51:48Z alexander.eichner@oracle.com $ */
 /** @file
  * IPRT - Expose a EFI variable store as a Virtual Filesystem.
  */
@@ -1051,7 +1051,7 @@ static DECLCALLBACK(int) rtEfiVarStoreDir_ReadDir(void *pvThis, PRTDIRENTRYEX pD
         {
             memcpy(&pDirEntry->szName[0], pszName, cbName);
             pDirEntry->szName[cbName] = '\0';
-            pDirEntry->cbName         = cbName;
+            pDirEntry->cbName         = (uint16_t)cbName;
             rc = rtEfiVarStore_QueryInfo(cbObject, fIsDir, &Time, &pDirEntry->Info, enmAddAttr);
             if (RT_SUCCESS(rc))
             {
